+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - SimpleDB: Use default server
 - Plugins: Method verifyVersion()
 - Plugins: Show driver plugins in the list of loaded plugins
+- Plugins: Show version of Adminer next to plugins not matching this version
 - Plugins: Remove edit-calendar, tinymce, json-column, pretty-json-column, translation, email-table, dump-php, master-slave
 
 ## Adminer 5.5.1 (released 2026-07-21)